+++ src/shared/lib/highlight-text.tsx
